feat(정책) : Repository 키값 타입 매핑 수정 (#163)

+++ src/main/java/com/project200/undabang/policy/repository/PolicyGroupMappingRepository.java
@@ -3,5 +3,5 @@ package com.project200.undabang.policy.repository;
 import com.project200.undabang.policy.entity.PolicyGroupMapping;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PolicyGroupMappingRepository extends JpaRepository<PolicyGroupMapping, Long> {
+public interface PolicyGroupMappingRepository extends JpaRepository<PolicyGroupMapping, Integer> {
 }
